requirements.txt
tensorflow==2.17.1
numpy==1.26.4
opencv-python==4.10.0.84
urllib3==2.2.3
matplotlib==3.8.0
pytesseract==0.3.13
pandas==2.2.2
scikit-learn==1.5.2
vertexai==1.71.1
google==2.0.3
google-auth==2.27.0
pandas
geopy
google-cloud-aiplatform
PyJWT
python-dotenv
flask